SonarPowerNoise (#2)
* SonarPowerNoise: WIP

* don't allow highpass > lowpass

* fix filter unit gain

The filter was computed as an amplitude gain map. This change computes energy
gains instead. This makes easier to ensure unit-gain and makes the alpha values
more in line with literature where the power spectrum is obeying the power law.
new alpha = old alpha / 2

Also fixes the gain of common_mode.

* allow stretch < 1.0, paramter range fixes

* rename lowpass/highpass to min_freq/max_freq

* remove torch.no_grad wrappers

* add previews

* static seed for previews

* README: SonarPowerNoise documentation